+++ euphonic/cli/utils.py
@@ -421,13 +421,31 @@ def _plot_label_kwargs(args: Namespace, default_xlabel: str = '',
 
 
 def _calc_modes_kwargs(args: Namespace) -> Dict[str, Any]:
-    """Collect arguments that can be passed to calculate_qpoint_phonon_modes()
+    """
+    Collect arguments that can be passed to
+    ForceConstants.calculate_qpoint_phonon_modes()
     """
     return dict(asr=args.asr, dipole_parameter=args.dipole_parameter,
                 use_c=args.use_c, n_threads=args.n_threads)
 
+def _brille_calc_modes_kwargs(args: Namespace) -> Dict[str, Any]:
+    """
+    Collect arguments that can be passed to
+    BrilleInterpolator.calculate_qpoint_phonon_modes()
+    """
+    if args.n_threads is None:
+        # Nothing specified, allow defaults
+        return dict()
+    else:
+        if args.n_threads > 1:
+            parallel = True
+        else:
+            parallel = False
+        return dict(useparallel=parallel, threads=args.n_threads)
 
-def _get_cli_parser(features: Collection[str] = {}
+
+def _get_cli_parser(features: Collection[str] = {},
+                    conflict_handler: str = 'error'
                     ) -> Tuple[ArgumentParser,
                                Dict[str, _ArgumentGroup]]:
     """Instantiate an ArgumentParser with appropriate argument groups
@@ -439,6 +457,9 @@ def _get_cli_parser(features: Collection[str] = {}
         Known keys: read-fc, read-modes, ins-weighting, pdos-weighting,
         powder, mp-grid, plotting, ebins, adaptive-broadening, q-e,
         map, btol, dipole-parameter-optimisation
+    conflict_handler
+        With default value ('error') parser will not allow arguments
+        to be re-defined. To allow this, set to 'resolve'.
 
     Returns
     -------
@@ -473,7 +494,8 @@ def _get_cli_parser(features: Collection[str] = {}
     _ins_choices = ('coherent',)
 
     parser = ArgumentParser(
-        formatter_class=ArgumentDefaultsHelpFormatter)
+        formatter_class=ArgumentDefaultsHelpFormatter,
+        conflict_handler=conflict_handler)
 
     # Set up groups; these are only displayed if used, so simplest to
     # instantiate them all now and guarantee consistent order/presence
@@ -485,7 +507,9 @@ def _get_cli_parser(features: Collection[str] = {}
                      'Force constants interpolation arguments'),
                     ('property', 'Property-calculation arguments'),
                     ('plotting', 'Plotting arguments'),
-                    ('performance', 'Performance-related arguments')
+                    ('performance', 'Performance-related arguments'),
+                    ('brille', 'Brille interpolation related arguments. '
+                               'Only applicable if Brille has been installed.')
                     ]
 
     sections = {label: parser.add_argument_group(doc)
@@ -761,6 +785,29 @@ def _get_cli_parser(features: Collection[str] = {}
                   'discontinuous segments of reciprocal space onto separate '
                   'subplots. This is specified as a multiple of the median '
                   'distance between q-points.'))
+
+    if 'brille' in features:
+        sections['brille'].add_argument(
+            '--use-brille', action='store_true',
+            help=('Use a BrilleInterpolator object to calculate phonon '
+                  'frequencies and eigenvectors instead of '
+                  'ForceConstants'))
+        sections['brille'].add_argument(
+            '--brille-grid-type', default='trellis',
+            choices=('trellis', 'mesh', 'nest'),
+            help=('Type of Brille grid to use, passed to the '
+                  '"grid_type" kwarg of '
+                  'BrilleInterpolator.from_force_constants'))
+        sections['brille'].add_argument(
+            '--brille-npts', type=int, default=5000,
+            help=('Approximate number of q-points to generate on the '
+                  'Brille grid, is passed to the "n_grid_points" kwarg '
+                  'of BrilleInterpolator.from_force_constants'))
+        sections['brille'].add_argument(
+            '--brille-npts-density', type=int, default=None,
+            help=('Approximate density of q-points to generate on the '
+                  'Brille grid, is passed to the "grid_density" kwarg '
+                  'of BrilleInterpolator.from_force_constants'))
 
     if 'dipole-parameter-optimisation' in features:
         parser.add_argument(